Добавил процедуры для бренда + SQL SERVER. Перед идентити.

+++ BmesRestApi/Messages/DataTransferObjects/Order/OrderDto.cs
@@ -6,9 +6,9 @@ namespace BmesRestApi.Messages.DataTransferObjects.Order
     public class OrderDto
     {
         public long Id { get; set; }
-        public decimal OrderTotal { get; set; }
-        public decimal OrderItemTotal { get; set; }
-        public decimal ShippingCharge { get; set; }
+        public int OrderTotal { get; set; }
+        public int OrderItemTotal { get; set; }
+        public int ShippingCharge { get; set; }
         public long CustomerId { get; set; }
         public int OrderStatus { get; set; }
         public long AddressId { get; set; }
